backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
tensorflow==2.20.0
numpy==2.2.1
Pillow==11.0.0
h5py==3.13.0
opencv-python
matplotlib
